server/requirements.txt
beautifulsoup4
httpx
loguru
openpyxl
python-telegram-bot[job-queue]
PyYAML
tortoise-orm
fastapi
uvicorn
